bindmount: Integrate tool which manages bind mount entries
This tool is intended to replace the functionality currently written in
resin-state-reset which creates the mount points in state partition before
carrying out the mount command (later in the mount systemd units).

When we bind mount a ro location, we want to keep the same structure in the
state partition and handle file and directories. For example if we want to
bind mount the file /etc/file this tool will make sure that
/mnt/state/root-overlay/etc/file exists and is a file, otherwise create an
empty one and carry on with bind mount. Similar to directories. It also warns
in case of file/directory shadowing.

Signed-off-by: Andrei Gherzan <andrei@resin.io>