balena: Don't inject TARGET_CC_ARCH as it should be included in CC
There is no need to include TARGET_CC_ARCH in LD/CC flags as they are already
included in CC (check bitbake.conf). This was harmless until rocko but with
rocko where gcc is compiled with default pie, it breaks the build because of:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=77529

What happens is that when compiling the objects for gcc_arm, the objects are
compiled with fPIC but later on, by injecting the TARGET_CC_ARCH which comes in
gcc command after the fPIC one, the latter gets deactivated. This makes
linking with -pie fail.

Signed-off-by: Andrei Gherzan <andrei@resin.io>