docker-disk: Fix compile error when not including any docker image
When building without the resin-supervisor (unmanaged production build for example)
then we do not include the supervisor image. In such a case, if we also do not bundle in
any container, the entry.sh script will not do any "docker pull" commands and then when sending
SIGTERM to the docker daemon it will shutdown immediately so that the subsequent wait() will
error out.

Change-type: patch
Changelog-entry: Fix docker-disk when not including any docker image
Signed-off-by: Florin Sarbu <florin@resin.io>