openvpn: Unify managed and unmanged images
This patch makes openvpn service to start based on the existence of
/etc/openvpn/openvpn.conf. This changes the existing behavior where this
service was only available in the connectable images (managed images) along
with the other resin specific bits like certificate, configuration etc. All
these extra bits are expected by the service to be in place before the service
starts so something else (in our case there will be a configuration service)
will do that prio to starting the openvpn service.

The openvpn service now starts with a bind mount dependency to make sure the
/etc/openvpn bind is in place.

As well, this change includes separating resin specific modifications to the
openvpn recipe in a different directory which makes updating the openvpn recipe
easier in the future. We use `files` for the changes we make for the resinOS
configuration and keep `openvpn` for the files coming from upstream. See:
http://cgit.openembedded.org/meta-openembedded/tree/meta-networking/recipes-support/openvpn

Signed-off-by: Andrei Gherzan <andrei@resin.io>