os-config: Initial integration of the OS configuration tool
This tool is responsible of configuring the OS based on os-config.json. It
communicaties with a configuration endpoint, gets the configuration and
restarts the affected services (as defined in the json mentioned above).

It also can run in cli mode which helps with configuring a completely
unconnected image, disconnect a connected image or switch in between different
resin instances (staging, production, etc.).

This tool's version includes support for generating device API keys as
well. The OS takes advantage of this functionality with a systemd service. This
service will replace the current resin-device-api-key.service.

Change-type: minor
Changelog-entry: Unify managed and unmanaged images and introduce os-config
Signed-off-by: Andrei Gherzan <andrei@resin.io>