update-resin-supervisor: Allow running update when no API endpoint is available
In unmanaged images, the API endpoint is not available while in resinOS
in container the supervisor is not preloaded. This means that when
running resinOS in container, the supervisor will get pulled in by the
updater only if the image is configured for resin.io connectivity (it
includes API endpoint). This makes resinOS in container not usable
(without tweaking) when running resinOS in container without config.json
resin.io configuration.

This change lets the update run even when no API_ENDPOINT is available
in which case the updater will just pull the version specified in
supervisor.conf.

Change-type: patch
Changelog-entry: Allow update resin spervisor when API endpoint is not available
Signed-off-by: Andrei Gherzan <andrei@resin.io>