dnsmasq: Split out resin changes from the upstream recipe
The main dnsmasq recipe was imported from the rocko branch of
meta-openembedded(1). After imported, we started to do modifications on the
imported recipe. This gets hard to maintain if we want to ever update the
package or when we want to point resin specific changes. To ease this, we split
out the resin changes in a separate bbappend and the new resources in a
separate directory.

Given that this imported recipe needs to work on multiple yocto versions we've
done the following change on the recipes:
1. Tweaked the calls to getVar. Some yocto versions require the expand argument
let's keep it to have this recipe working with more yocto versions.
2. Replace filter by contains calls. This is because the filter is not
available in some resin supported yocto versions (ex. morty).

(1) https://github.com/openembedded/meta-openembedded/tree/rocko/meta-networking/recipes-support/dnsmasq

Change-type:patch
Changelog-entry: Split resin specific changes in dnsmasq in a separate bbappend
Signed-off-by: Andrei Gherzan <andrei@resin.io>