docker-resin-supervisor-disk: Pass self-signed CA data to the Supervisor and for Docker registry trust.
This patch ensures that any self-signed CA that has been passed to
the system is decoded and then stored for the specified registry.
It also sets the appropriate envvars for the Supervisor so that
may also read the CA and therefore trust it.

Change-type: patch
Changelog-entry: Pass self-signed CA data to the Supervisor and for Docker registry trust.
Connects-to: #1044
Signed-off-by: Heds Simons <heds@whaleway.net>