resin-extra-udev-rules: Add teensy udev rules
[Issue #1063]

When a Teensy micro controller is connected to a resin device the
ModemManger attempts to use the Teensy as a modem. This makes serial
communication with the Teensy from within an application container
unreliable.

These changes add a recipe called `resin-extra-udev-rules` which copy
the Teensy rules into the udev rules folder which prevent ModemManger
from attempting to use a Teensy as a modem.

Change-type: patch
Changelog-entry: Adds Teensy udev rules
Signed-off-by: Henry Miskin <hpgmiskin@gmail.com>