timeinit: Add service for updating system clock based on RTC
This service starts before timeinit-timestamp and systemd's timesyncd. It only
starts if an rtc is available based on the existence of /dev/rtc.

The ordering based on timeinit-timestamp service is because the timestamp
service doesn't update the clock atomically and if they start in parallel, we
could end up with the system clock set by the timestamp even if the RTC has a
newer time.

The actual ordering, before or after, doesn't really matter because timestamp
will only update if the current system clock is before the timestamp. We use
though a "Before" systemd unit option, so the RTC service will start before the
timestamp one.

Change-type: patch
Changelog-entry: Use system clock at boot based on RTC when available
Signed-off-by: Andrei Gherzan <andrei@resin.io>