timeinit: Fix systemd service for timeinit-timestamp.service
First of all timesyncd doesn't exist. The correct service is systemd-timesyncd.

Along with this change, we removed the balena dependency because it was useless
anyway. Balena starts in multi-user target while timesyncd starts in
time-sync.target which is enabled before multi-user. So as long as the
timestamp service starts before timesyncd, it will start before balena as well.

Change-type: patch
Changelog-entry: Fix typo in timeinit-timestamp service
Signed-off-by: Andrei Gherzan <andrei@resin.io>