chrony: Add recipe to resinOS
This patch enables the chrony package in resinOS.

The chrony recipe was added in morty. So add the base recipe
in previous yocto versions with a minor modification.
chrony_%.bbappend is kept in meta-resin-common to keep
resinOS specific configuration in one place.

systemd-timesyncd will be replaced by chrony as chrony is more
feature rich.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>