busybox: Add patch to install  in /usr/bin
Since sumo the build system checks if the post install scripts fail and the
rootfs is ro in which case it will error out and break the build.

coreutils and busybox provide the `link` utility and the build system uses
update-alternatives to avoid conflicts. The issue is that coreutils installs
the link in /usr/bin while busybox in /bin which makes update-alternative
complain and fail at rootfs generation step and break the build due to the build
system check described above.

We fix this issue but patching busybox to install `link` in /usr/bin as
coreutils does.

Signed-off-by: Andrei Gherzan <andrei@resin.io>