resin-ntp-config: Fix systemd service dependencies since switch to chrony
https://github.com/resin-os/meta-resin/pull/1077 introduced the swtich to
chrony from systemd's timesyncd. resin-ntp-config was modified to now use
chronyc when setting the NTP servers from config.json so the systemd service
needs now to reflect that this change.

As well the recipe now depends at runtime on chronyc (as it runs this client)
and on chrony (as it depends on the daemon's service).

Signed-off-by: Andrei Gherzan <andrei@resin.io>