NetworkManager: Add systemd service dependency to chronyd
NetworkManager uses a dispatcher script which manages NTP servers from DHCP but
in order for this to work properly we need to make sure chronyd is running.
Ensure this by having the right dependency in NM's systemd service.

As well the recipe now depends at runtime on chronyc (as it runs this client in
a dispatcher script) and on chrony (as it depends on the daemon's service).

Signed-off-by: Andrei Gherzan <andrei@resin.io>