networkmanager: Set DHCP timeout to infinity for IPv4
We switch the DHCP IPv4 request timeout to infinity so that we circumvent
situations when the DHCP server times out for long periods of time during
which the lease may also expire and thus when the DHCP server is restored
NetworkManager will not retry at that point as the timeout interval may be
long expired.

We set the timeout to 2147483647 which is basically MAXINT32, but NM will
efectively never timeout in fact.

Change-type: patch
Changelog-entry: Set DHCP timeout to infinity for IPv4
Signed-off-by: Florin Sarbu <florin@resin.io>