kernel-resin.bbclass: Add CONFIG_PANIC_TIMEOUT
Currently if the kernel panics, it is up to the device kernel
configuration if it timesout and reboots on a kernel panic.

Also, the watchdog if present does not reboot a device that is in
a panic state as the kernel watchdog driver keeps touching it.

Enforce a reboot on kernel panic from meta-resin.

Change-type: patch
Changelog-entry: Reboot device if there is a kernel panic
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>