resin-image: Fix fingerprints
Since rocko update, this commit ended up in poky:

	commit bb81ff1343d8d6213f0418fd18b10312c96c4617
	Author: Richard Purdie <richard.purdie@linuxfoundation.org>
	Date:   Fri Aug 11 08:10:33 2017 +0100
	    image-prelink: Disable for musl images

This switched the use of `+=` to `_append` when adding the prelink hooks to
IMAGE_PREPROCESS_COMMAND. Because we are using `+=` for adding our fingerprints
hook, we end up having the prelink hooks after the fingerprints one (append is
evaluated after parsing - see
https://www.yoctoproject.org/docs/1.6/bitbake-user-manual/bitbake-user-manual.html#appending-and-prepending-override-style-syntax).
This results in wrong artifacts when prelink modifies the binaries - rootfs
ends up with wrong fingerprints. We fix this be using `_append` as well.

Change-type: patch
Changelog-entry: Fix rootfs fingerprints on poky rocko or newer
Signed-off-by: Andre Gherzan <andrei@resin.io>