cargo,rust: Update Morty packages with the versions in meta-rust pyro
We have package in resinOS (for example os-config) which depend on a rust
version >= 1.16. Morty comes with an older version with which crates like log
can't compile (see https://github.com/rust-lang-nursery/log/issues/252). The
newer meta-rust versions (from pyro on) already include versions which satisfy
this requirement.

As an intermediate step we import rust recipes (and cargo recipes) from
meta-rust pyro which will upgrade rust to 1.20 (and cargo to 0.21). The exact
revision from which this import was done is (meta-rust):
    commit 94280a636c9a6dd729f9eb3ed88fe85b11ee8ffa
    Merge: f6d973c 4c764c3
    Author: Derek Straka <derek@asterius.io>
    Date:   Tue Oct 10 12:06:43 2017 -0400

As well, we refresh the `cargo` bbclass and include additional bbclasses used
by these new recipes.

Change-type: patch
Changelog-entry: Update rust to 1.20 and cargo to 0.21.0 in Morty
Signed-off-by: Andrei Gherzan <andrei@resin.io>