resin-mounts-flasher: Simplify boot partition mount(s) in the case of flasher
For the sake of keeping the config.json in the same location in a way that it
would be rw but also volatile (so that we can use the same flasher for multiple
provisioning), we were using a combination of bind mounts and mounts that
brings quite some overhead. It's much easier to just mount the boot partition
completely ro and set up a config.json in /tmp where all services are pointed
as per resin-vars logic. In this way we have the same `volatile` requirements
achieved and also the scripts don't need to behave differently on
flashers/normal images. The only flasher specific logic (which detects the
image type and sets the correct config.json location) is in resin-vars.

Signed-off-by: Andrei Gherzan <andrei@resin.io>