resin-mounts: Add chrony drift file bindmount
Chrony stores time drift in a file occasionally (and at least
once an hour). Bind mount this file so it persists on reboot.
The timestamp of this file is useful on first boot as chrony reads
that and sets the system time to the timestamp if it happens to be
the latest time

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>