resin-image: Handle absolute paths in RESIN_BOOT_PARTITION_FILES
Currently the device-type.json is considered relative to deploy
directory. Which works fine as long as the deploy directory is in a
known location. If user takes advantage of custom TMPDIR and/or
DEPLOY_DIR, this reference brakes while finding the device-type.json.

This patch fixes this by allowing absolute paths in
RESIN_BOOT_PARTITION_FILES for entries like device-type.json which have
a known location in the git repository.

Fixes https://github.com/resin-os/meta-resin/issues/1121

Change-type: patch
Changelog-entries: Fix build when using custom TMPDIR and/or DEPLOY_DIR
Signed-off-by: Andrei Gherzan <andrei@resin.io>