resin-device-register: Add pin call after we register the device
This package is used to pin devices when these are preloaded. The
mechanism is present in the supervisor, but because pinning needs to
happen when provisioning, and resin-image-flasher does that
independently, we require this package to take care of pinnig in the
flasher. This package is configuared via a
file(/etc/resin-device-pinning.conf) which if created should contain
RELEASE_ID and represents where to pin the device. This file is created
by resin-preload. This is a temporary fix and will be made redundant
once the provisiong process is unified.

Changelog-entry: Allow flasher types to pin preloaded devices
Change-type: minor
Signed-off-by: Theodor Gherzan <theodor@resin.io>