resin-device-uuid: Avoid race with os-config-devicekey
Avoid racing with the os-config-devicekey service by adding an order
directive in the service. As well we move this service in basic target
to assume uuid in later targets (like multi-user).

The problem at hand is that os-config-devicekey reads and write the
config.json which races with the fact that uuid generation service
writes as well to the same config file. If this happens in between
os-config's read and write, we overwrite the changes done by the uuid
service.

Fixes #1159

Change-type: patch
Changelog-entry: Avoid os-config-devicekey / uuid service race
Signed-off-by: Andrei Gherzan <andrei@resin.io>