hostapp-update-hooks: Update grub hooks to prepare to load from root
The kernel will be located in the root partition in /boot folder.
This requires an update to grub.cfg in the integration layers
to search for resin-rootA/B and set that as the root for searching
the kernels.

This updates grub.cfg to have multiple mentions of resin-rootA/B
both of which need to be switched. Update hostapp-update hooks to
replace all instances of resin-rootA/B instead of just the single
mention

Change-type: minor
Changelog-entry: Update grub hooks to prepare to load kernel from root
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>