networkmanager: Replace NM's DHCP request option "Client indentifier"
This patch replaces the "Client identifier" option from
DHCP packets sent by NM to resemble the way udhcpc does it.

The reason is that the Cisco RV325 router with fw older than
v1.2.1.14 doesn't reply to DHCP requests or discovery packages
if the "Client identifier" option is formatted in NM's style.

Even though this is a bug in the router's fw, we are going to
fix it here because we encountered clients that couldn't get
an IP address on these routers and we want to avoid that

Change-type: patch
Changelog-entry: Replace NM's DHCP request option "Client indentifier" with udhcpc style option
Signed-off-by: Sebastian Panceac <sebastian@balena.io>