recipes-containers: Increase healthcheck timeout to 180s
The original 60s might be slow for some devices where balena
is busy with a task and the hardware is slow for various
reasons (ie. Pi1/Zero). In those cases even if the device
is functioning correctly, the healthcheck will assess the
device as unhealthy and restart balena and/or the supervisor.

Change-type: minor
Signed-off-by: Gergely Imreh <gergely@balena.io>