resin-mounts: Increase timeout for the filesystem label
The label is actually a udev symlink, but we're keeping
this naming to be consistent across the file.

The error seen is that the balena-host service wasn't
able to start because of a mnt-sysroot-active.service dependency:
resin-partition-mounter[440]: ERROR: Timeout while waiting
for filesystem label to appear

By increasing this timeout, the filesystem symlink should
already exist, at the time the mnt-sysroot-active service is
looking for this dependency.

Change-type: minor
Changelog-entry: Increase timeout for filesystem label
Signed-off-by: Vicentiu Galanopulo <vicentiu@resin.io>