networkmanager: Add chronyc online in dispatcher script
If an interface comes up, tell chronyc that an interface is online
to make it start synchronizing time.

We purposefully don't add the chronyc offline command in the down
path of this script. There can be multiple network interfaces and we
don't want one of them to go down and take chrony down with it.

Changetype: minor
Changelog-entry: Tell chrony that an interface is now online via networkmanager
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>