modemmanager: Respect FILESEXTRAPATHS when adding bbappend files
The bbappend uses a local `resin-files` directory for additional file
entries but some of these files ended up in `files`. This works well if
the recipe used is the one in the same directory as the `files`
subdirectory will be used by default. The problem shows when the recipe
is not in the same directoy as bitbake will not be able to find needed
bbappended files.

Change-type: minor
Changelog-entry: Fix modemmanager bbappend files
Signed-off-by: Andrei Gherzan <andrei@balena.io>