networkmanager: Remove tun devices from unmanaged list
Under the current configuration we instruct NetworkManager through its
configuration to completely ignore all the tun devices. This was
enforced as an extra precation as we run an openvpn client which
connects to the balena-cloud backend. This client creates, on the host,
the interface `resin-tun` currently (rename pending). We want to loosen
up this restriction to only unmanage this specific device - `resin-tun`
- so that users can take advantage of NetworkManager managing other tun
devices.

Change-type: minor
Changelog-entry: Configure NetworkManager to only ignore our vpn interface but manage other tun devices
Signed-off-by: Andrei Gherzan <andrei@balena.io>