resin-u-boot.bbclass: Add bootcount variables to u-boot environment
Currently, we rely on u-boots bootcount framework and driver to
increment the bootcount. That doesn't scale easily across all devices.
We'd like to handle bootcount increments in env_resin.h.

In preparation for handling bootcount increment in env_resin.h.
Define some variables that we will use in env_resin.h

OS_BOOTCOUNT_SKIP allows devices to disable bootcounting in u-boot.
This is for some devices that might not reliably support fatwrite in
their bootloaders.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>