image_types_resin.bbclass: Fix img dependency on rootfs
There can be cases where a image type gets invalidated (due to a
variable change or so) and the dependency of it not. For example the img
type can get invalidated while the hostapp ext3 type doesn't. The img
generatiton script consumes/uses the rootfs generated by the hostapp
type but, it references the artifact based on IMAGE_NAME variable which
includes a time stamp. See bitbake.conf:

IMAGE_NAME = "${IMAGE_BASENAME}-${MACHINE}-${DATETIME}"

When that happens, the new timestamp will change and img will fail to
find the old artifact of hostapp. Change this behaviour by using the
symlink reference IMAGE_LINK_NAME which doesn't include a timestamp.

Change-type: patch
Changelog-entry: Fix img to rootfs dependency when img is invalidated
Signed-off-by: Andrei Gherzan <andrei@balena.io>