os-release: Fix not deterministic metadata
When meta-balena and/or balena-board repositories change the revision,
building os-release from sstate breaks with something similar to:

  ERROR: When reparsing
  /build/../layers/poky/meta/recipes-core/os-release/os-release.bb.do_compile,
  the basehash value changed from f4bc42cbadaf15b0a62c64f697adcd47 to
  5f742349ff06f250941b11dca848ed60. The metadata is not deterministic and
  this needs to be fixed.

This is because these revisions are part of the taskhash and make it
change when an already populated sstate is used that was created with a
different repository revision variable value. We fix this by moving the
generationg of this additional information (revisions of balena repos)
from os-release to the image.

This commit also restructures the code a bit.

Fixes #1421

Signed-off-by: Andrei Gherzan <andrei@balena.io>