cargo/sumo: Make bbappend version independent
The current bbappend, which only copies the lock file, is made version
specific. This commit uses a version wildcard which will unlock updating
the cargo package from meta-rust while still keeping the fix for cargo
versions that still use the lock file. This lock file, in later versions
is not used anymore:

See
https://github.com/rust-lang/cargo/commit/5c9665f41c6b4d3b99d3b9f8b48a286f5f154692

We will be unifying the rust compiler while updating it to a newer
version (1.33). This version doesn't need a lock file anymore as
mentioned above.

Changelog-entry: Make cargo bbappend version independent for sumo
Signed-off-by: Andrei Gherzan <andrei@balena.io>