systemd: Fixes for sysroot symlinks creation
When we set the udev rule we use IMPORT{program} to know what symlinks
to create in the sysroot directory. The problem is that we don't use
absolute paths and from the documentation udev uses /usr/lib/udev as the
default path while we install the invoked script in /lib/udev.

```
  Execute an external program specified as the assigned value. If no
  absolute path is given, the program is expected to live in
  /usr/lib/udev; otherwise, the absolute path must be specified.
```

This PR changes the udev rule to use an absolute path for the
resin_update_state_probe script. Also it fixes some typos in the script
itself.

Change-type: patch
Changelog-entry: Fixes for sysroot symlinks creation
Signed-off-by: Andrei Gherzan <andrei@balena.io>