uboot: Add support for unified kernel cmdline arguments
This adds set_os_cmdline which defines os_cmdline so that BSPs can
import it in their bootargs and pass it to kernel accordingly. Also, as
part of os_cmdline, there is an argument added when the booted image is
a flasher. This can be used in the OS for various flasher specific
checks.

Change-type: minor
Changelog-entry: Add uboot support for unified kernel cmdline arguments
Signed-off-by: Andrei Gherzan <andrei@balena.io>