balena-linux-firmware: Firmware clean various improvements
This adds an easier to debug logging that shows both removed and kept
firmwares. For example:

 Cleaning iwlwifi firmware version 8000C to minimum API version 34.
   Removing iwlwifi-8000C-13.ucode.
   Removing iwlwifi-8000C-22.ucode.
   Removing iwlwifi-8000C-21.ucode.
   Removing iwlwifi-8000C-31.ucode.
   Removing iwlwifi-8000C-16.ucode.
   Removing iwlwifi-8000C-27.ucode.
   Leaving iwlwifi-8000C-34.ucode in place.
   Leaving iwlwifi-8000C-36.ucode in place.

Also, this commit improves the version comparison which was comparing
strings and not integers. Now we try integer comparison and fallback to
string when parse fails.

Finally, the task's vardeps expands varflags to correctly retrigger on
change.

Fixes #1527

Change-type: minor
Changelog-entry: Improve logging and version comparison in linux-firmware cleanup class
Signed-off-by: Andrei Gherzan <andrei@balena.io>