busybox: Use glibc's resolver as opposed to the internal one
Since 1.29, busybox switched to an internal implementation of the
resolver based on a feature config, NSLOOKUP_BIG. This is enabled by
default and it's meant to be musl compatible. In BalenaOS we use glibc
and we rely on it for being able to resolve names using NSS modules
(libmdns).

Change-type: patch
Changelog-entry: Fix busybox nslookup mdns lookups
Signed-off-by: Andrei Gherzan <andrei@balena.io>