bluez5: Persist storage directory over reboots
Currently, bluez's storage data is set to /var/lib/bluetooth which
in turn is a tmpfs location. We want this location persistent so we can
save paired devices over reboot. We do that by adding the corresponding
bind mount to the state partition and setting bluez to depend on this
mount unit.

Fixes #1544

Change-type: minor
Changelog-entry: Persist bluetooth storage data over reboots
Signed-off-by: Andrei Gherzan <andrei@balena.io>