image-resin.bbclass: Fix to be able to use deprecated layers
Fixes #1564

This bbclass uses the common layer path advertised by BBLAYERS to find
the paths to different resources in board repository. For example
machine json - to get the slug and advertise it in os-release. After
renaming the common layer (resin to balena) and because we support old
references to the old common layer name (by including a dummy
deprecation layer), we need to make sure this class can cope now with
both cases.

Change-type: patch
Changelog-entry: Fix image-resin.bbclass to be able to use deprecated layers
Signed-off-by: Andrei Gherzan <andrei@balena.io>