u-boot: Add support for balenaos_uboot config fragments
We'd like to enable some config options for all devices via meta-balena

The support is mainline since warrior. Lets copy it in meta-balena
to get it in the previous layers as well.

We need CMD_SETEXPR to be enabled for rollbacks to work. So enable it
via the config fragment

Change-type: minor
Changelog-entry: Devices using u-boot. Enable CONFIG_CMD_SETEXPR for all devices. Required for rollbacks to work
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>