meta-balena-common: Add balena-image-migrator-initramfs target
This is based on the resin-image-initramfs. We strip out scripts that
are only needed for balenaOS and add the migrator package and some
other utilities needed by the migrator. e.g. sfdisk, dosfstools, mke2fs
badblocks.

You can now run MACHINE=xyz bitbake balena-image-migrator-initramfs
This will generate a kernel with the migrator initramfs built in.

Change-type: minor
Changelog-entry: Add new target that can be used to generate a migrator initramfs
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>