docker-disk: Reduce data partition size from 1G to 192M
The data partition contains the supervisor which is only about 61M on
the pi3. We compress the data partition later on so don't notice these
zeros. But lets reduce the size of the data partition to eat less
space whenever an uncompressed image is used anywhere. e.g. when
extracting the zip to dd or using balena-migrate etc.

The partition expands on first boot to fill the rest of the disk
anyways.

Change-type: minor
Changelog-entry: Reduce data partition size from 1G to 192M
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>