docker-disk: Pull the hello-world image into the data partition
This image will be used for a simple health-check for the balenaEngine
daemon.

We run the build on amd64 machines. During the build, we cannot run a
simple docker pull arm32v5/hello-world on amd64 machines. This returns
a no matching manifest error.

We work around this issue by using the experimental --platform flag to
fetch the different architecture images. That flag requires a bump to
18.06 dind image.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>