docker-disk: Save hello-world image in a tarball
For the new hello-world healthcheck, devices need the hello-world image
We can't rely on internet connectivity and access to dockerhub registry
for devices behind firewalls, on-prem etc. Therefore, lets include the
hello-world image tarball in the OS itself.

The image will be loaded by the balena daemon when the service starts.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>