resin-supervisor: Expose container ID
In some usecases, it's very helpful for supervisor to know its container ID.
For instance, when supervisor leaves local mode, it removes all the engine objects
that didn't not exist before entering local mode.
Supervisor needs to know its own container ID to not delete itself.

Using info in `/proc` files seems to be engine implementation dependant
and may not work in some cases. So this way seems to be the most reliable to expose
container ID to the application.

Change-type: minor
Signed-off-by: Roman Mazur <roman@balena.io>