hostapp-update: Update for hooks 2.0
With hooks 2.0 we now have multiple pre/post migration hooks. The
trickiest one are the pre hooks. These are hooks that have to be
downloaded in the update package and run in the old OS environment.
Pre-hooks have to be written quite carefully keeping in mind the
current working directory is in the sysroot inactive partition and
not the usual /etc/hostapp-update-hooks.d that hooks expect to be in.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>