balena: Double watchdog timeout to accomodate for the new health-check
We have now added a balena run --rm hello-world healthcheck. On a pi3,
this takes about 3-4 seconds. On a pi1, this can take much longer.

This also depends on the application load too.

Double to timeout to accomodate for the improved health-check

Changelog-entry: Increase balenaEngine service watchdog timeout to 360seconds
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>