kernel-modules-headers: Pull in gen_mod_headers into meta-balena
We have this one script in an external repo for not much reason.
https://github.com/balena-os/module-headers

It is awkward to patch and then update the recipe. Also, we would like
to test the generated tarballs via a recipe in meta-balena. So it makes
sense to have the generation script in the same repo as well so that
updates can update the tarball generation + testing in one go.

Pull in v0.0.20 1098563

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>