systemd: Vacuum journald on boot
We have found a rare corner case bug where the journal bloats beyond
its limit and fills the state partition. Triggering a vacuum on reboot
helps a bit in case the device is restarted to recover its function.

Fixes #1423

Change-type: patch
Changelog-entry: Workaround for a cornercase bug in PersistentLogging where journalctl filled the state partition. Vacuum the journal on boot.
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>