initrdscripts/warrior: Add console=null kernel cmdline workaround
When `console=null` is passed in the kernel cmdline for production
images, the system doesn't boot. Traced to initramfs not starting any
process such as udev with the right file descripters for 0,1,2

Add workaround for that issue. This got exposed with a systemd bump.
More details here https://github.com/systemd/systemd/issues/13332

Change-type: patch
Changelog-entry: Fix a hang in initramfs for warrior production images
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>