balena-rollback: Fix service dependency
PR #1441 changed mnt-sysroot-inactive to an automount. But there is no
way to easily add a udev dependency to the automount. As a result,
when rollbacks tries to access the inactive partition, it fails.
Let's add systemd-udev-settle.service to the rollback services.

This will only delay the first boot after a HUP which is reasonable.

Change-type: patch
Changelog-entry: Fix for rollbacks where the inactive partition mount was unavailable when altboot triggered
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>