initrdscripts: Improve workaround
We would just check for the existence of console=null in kernel cmdline

A better way would be to check the presence of a valid symlink for fd0,
fd1 and fd2. And assign them to /dev/null if unavailable.

Change-type: patch
Changelog-entry: A small fix in initramfs when /dev/console is invalid due to whatever reason
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>