kernel-resin.bbclass: Build in FAT fs related modules
balenaOS uses FAT as a fs type for the boot/first partition. This is
currently hardcoded so let's have the related kernel configs built in
the kernel image. In this way we don't have to handle kernel modules in
the initramfs (when needed).

Change-type: patch
Changelog-entry: Switch to built-in FAT kernel configs
Signed-off-by: Andrei Gherzan <andrei@balena.io>