systemd: Detect label conflicts and ignore when setting active/inactive symlinks
It can be the case where multiple "resin-root*" filesystem labels are
available at runtime. For example when the OS handles balenaOS disk copy
operations. In this case, udev needs to ignore the labels that have
correspondent UUIDs not part of the stored ones at first boot when
creating the active/inactive symlinks.

Signed-off-by: Andrei Gherzan <andrei@balena.io>