resin-supervisor: Ask balenaEngine to restart an unhealthy
Currently balenaEngine knows a supervisor is unhealthy but we wait for
healthdog to check the status and then restart supervisor.

Lets ask balenaEngine to restart the supervisor if it knows the
supervisor is unhealthy.

Healthdog is now restricted to checking presence of the supervisor
image itself.

Fixes #1650

Changelog-entry: Modify supervisor healthcheck. balenaEngine will restart an unhealthy supervisor if needed
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>