dropbear: Ignore conflict from openssh package
Zeus improved dependency checks and caught this inconsistency.
We have switched to using openssh in balenaOS. But still need some
dropbear binaries to migrate the old keys as part of the hooks.

As a result we install dropbear and openssh both. But Yocto doesn't
like that. Append the recipe and ask Yocto to ignore the conflict.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>