resin-image: Remove ssh from IMAGE_FEATURES
We have switched to using openssh in balenaOS. But still need some
dropbear binaries to migrate the old keys as part of the hooks.

As a result we install dropbear and openssh both. But Yocto doesn't
like that. We have not explicitly asked Yocto to install dropbear and
openssh via packagegroup-resin-connectivity. Remove from IMAGE_FEATURES

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>