balenaEngine: Namespace the hello-world healthcheck image
Fixes #1630

Need to make the script go in the background in ExecStartPost=.
Otherwise, the service status never gets to active/running resulting
it in remaining stuck in an endless loop.

If the health-check load fails for whatever reason, the subsequent
engine healthcheck will fail retriggering the healthcheck load.

Change-type: minor
Changelog-entry: Namespace the hello-world healthcheck image
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>