resin-init-flasher: Tweak detection of internal media
It is better to use fdisk to check that an internal media
is a suitable block device rather than using the shell
internal block testing feature. This is because there is
at least a case when the kernel lists an ASMT 2115 controller
as /dev/sda but that is not an actual drive so it cannot be
written to and the flashing procedure will error out.

Change-type: patch
Changelog-entry: Tweak how the flasher asserts that internal media is valid for being installed balena OS on
Signed-off-by: Florin Sarbu <florin@balena.io>