Increase the NTP poll rate
This commit increases the default NTP poll rate to around 4.5 hours.

BalenaOS uses chronyd for time synchronization, and it allows to specify
a minpoll and maxpoll values per server with a power of two number of
seconds for the minimum and maximum polling time respectively.

With those constraints, the change set both limits to 2^14s (4.55h) for
all servers.

An architectural decision has been made not to make this configurable.

Fixes #1780.

Change-type: patch
Changelog-entry: Increase NTP polling time to around 4.5 hours.
Signed-off-by: Alex Gonzalez <alexg@balena.io>