resolv.conf: Increase DNS client's timeout to 15 seconds
The DNS clients (applications) resolver libraries use the timeout value in
/etc/resolv.conf to set the time between DNS attempts. The default is 5
secs which for slow networks like cellular mean lots of DNS requests on
a bandwidth sensitive channel.

This change modifies the default to 15 secs. This timeout only applies
when DNS servers are unresponsive so it will not affect the normal
functionality.

Fixes #1800

Change-type: patch
Changelog-entry: Increase DNS clients timeout to 15 seconds
Signed-off-by: Alex Gonzalez <alexg@balena.io>