balena-engine: Patch out hardcoded CC for cross-compile
Port of meta-virtualization/docker commit:
https://git.yoctoproject.org/cgit/cgit.cgi/meta-virtualization/commit/recipes-containers/docker/docker-ce_git.bb?id=e7787cb01be91d3798594687a475d5b085f29ede

Since commit applied in moby [61a3285 Support cross-compile for arm]
it hardcoded var-CC to support cross-compile for arm

Correct it with "${parameter:-word}" format, it is helpful for user
define toolchains

(Use Default Values. If parameter is unset or null, the expansion of
word is substituted. Otherwise, the value of parameter is substituted.)

https://github.com/moby/moby/commit/61a3285864d3f1b489f48f765b61b2c7bd300372

Signed-off-by: Robert Günzler <robertg@balena.io>