u-boot: Use UUIDs for the root partition instead of PARTUUIDs.
We are regenerating filesystem UUIDs to make them unique, so it is
sensible to also use filesystem UUIDs to boot instead of partition UUIDS
which will not be unique.

This commit also adds filesystem UUID support to the resin-u-boot class.
It's done on the class and not as a configuration fragment as the change
needs to apply also to BSP specific u-boots, like u-boot-tegra.

Signed-off-by: Alex Gonzalez <alexg@balena.io>