update-resin-supervisor: Allow to run any image locally available
This commit removes the requirement of only running an image that has been
successfully pulled. If the specified image is available locally, let's
use it.

The process to run a local image would be:

* Create a tarball with "balena save"
* Load it locally with "balena load -i <tarball>"
* Tag it with "balena tag <image id> <repo>:<tag>
* Update resin-supervisor with:
```
update-resin-supervisor -i <repo> -i <tag>
```

This provides a way to perform local offline updates as requested in:
https://github.com/balena-os/meta-balena/pull/1827

Changelog-entry: Allow local resin-supervisor image updates
Changelog-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>