balena-host: Allow to provide environment variables.
This allows for example to configure a DOCKER_TMPDIR environment variable
that will be used to set the temporary download location for the balena
host engine.

This works around the "no space left on device" problem on offline
hostOS updates requested in:
https://github.com/balena-os/meta-balena/pull/1827#issuecomment-581858707

Changelog-entry: Allow to configure the environment for balena host.
Changelog-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>