update-resin-supervisor: try to create & use delta images
rather than relying on pulling full images from DockerHub, we
can use supervisor apps in balenaCloud to generate and apply deltas.

This change is compatible with both old-style "${image}:${version}"
responses, as well as the (soon-to-come) new-style
"${registry-address}:latest" responses.

Finally, the code used in this change is essentially equivalent to what
is used in HUP for OS deltas.

[ ]: old-style "$image:$tag" -> new "$location:latest"
[ ]: new "$location:latest" -> new "$location:latest"
[ ]: update cycle "$location:latest" -> same "$location:latest"

Connects-to: https://github.com/balena-io/balena-io/pull/2177
Depends-on: https://github.com/balena-io/balena-api/pull/2786
Change-type: patch
Signed-off-by: Matthew McGinn <matthew@balena.io>