systemd-zram-swap: Add compressed memory swap support
This enables a compressed RAM swap of 25% of total system memory or 1GB,
whichever smallest.

Upstream Yocto has a zram implementation that's broken in warrior, and a
new implementation introduced in Zeus that does not work in BalenaOS as
we run two udevs, one in the initramfs and one in the main OS. The
mkswap needs to happen in the initramfs udev otherwise the udev database
is not updated with the swap device and the zram0 device is not detected
by systemd.

This implementation is simpler than the one upstream and common to all
the supported Yocto versions. It uses a udev rule in the initramfs that
creates the swap drive, and a swap unit in the main OS that enables it.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>