balena-engine: Compress binary
As the root filesystem size limit is being reached, compressing the engine
provides around 30M of space saving with a penalty on boot time.

The smaller binary loads quicker so the in-memory decompression is
balanced out with faster loading times, specially in devices with slower
storage.

On a Intel NUC the boot time penalty is around 5s.

Signed-off-by: Alex Gonzalez <alexg@balena.io>