resin-image: Break down the rootfs image size calculation
Replace the IMAGE_ROOTFS_SIZE magic number with a calculation based on
the partition sizes on the balena image classes.

The rootfs size is calculated with the assumption of a total BalenaOS
size of 700M. This includes boot, state, rootA and rootB partitions. The
data partition will then grow to occupy the rest of the storage media.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>