image_types_resin: hostapp-ext4: Fail if not enough space on filesystem
Before creating the hostapp-ext4 filesystem, check that there will be enough
space in the disk for it to fit.

The ext4 filesystem reserves a 5% of space by default and also contains
a journal, so substract those from the image size.

Also, substract an extra 5% which is empirically needed by the
hostapp-update process.

Changelog-entry: Stop build if hostOS update file would not fit in rootfs partition
Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>