resin-ntp-config: merge 'burst' command with 'add server' line
When the user supplies an additional NTP server source which is a
pool URL the 'burst' command may fail. This occurs when the pool
URL resolves to a different IP addresses for the 'add server' and
'burst' commands.

To avoid this issue we can combine the burst functionality into the
'add server' command by using the 'iburst' option. Although this
option is not documented by the chronyc man page it has been present
since v1.25 released in 2011.

This fix has been tested via the balenaOS (2.51.1+rev1) command line
running on a RPi3.

Change-type: patch
Connects-to: #1903
Signed-off-by: Mark Corbin <mark@balena.io>