resin-supervisor: Create required directories before launch
On commit a4ce26caadabcb1e87d944d78218cc32c579914e the supervisor moved
from using --volume to using --mount to avoid the implicit creation of
directories instead of files.

However, in the case where the mount referred to a directory, these have
to exist in the rootfs beforehand as --mount will not create them.

This commit checks for the existence of the /var/log/supervisor-log
directory and creates it if required.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>