Respect balenaRootCA system-wide
We allow the user to specify a custom CA in the .balenaRootCA key
of config.json but at this moment each tool has to implement support
if it wants to use it. This commit adds it to the system-wide CA bundle
so that the CA is respected everywhere.

Fixes #1398

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>